// pop-web/src/app/shared/openapi/model/lifestyle-create.ts
/**
 * POP API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CodedConcept } from './coded-concept';
import { Measure } from './measure';


export interface LifestyleCreate { 
    externalSource?: string | null;
    externalSourceId?: string | null;
    /**
     * Indicates the case of the patient who\'s lifestyle is assesed
     */
    caseId: string;
    /**
     * Clinically-relevant date at which the patient\'s lifetyle was assessed and recorded.
     */
    date: string;
    smokingStatus?: CodedConcept | null;
    smokingPackyears?: number | null;
    smokingQuited?: Measure | null;
    alcoholConsumption?: CodedConcept | null;
    nightSleep?: Measure | null;
    recreationalDrugs?: Array<CodedConcept> | null;
    exposures?: Array<CodedConcept> | null;
}